move shared components and update routes
