+++ src/app/list/list.component.spec.ts
@@ -3,8 +3,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 // COMPONENTS
 import { ListComponent } from './list.component';
-import { HeaderComponent } from '../shared/header.component';
-import { FooterComponent } from '../shared/footer.component';
+import { HeaderComponent } from '../shared/header/header.component';
+import { FooterComponent } from '../shared/footer/footer.component';
 
 describe('ListComponent', () => {
   let component: ListComponent;
