+++ src/app/shared/footer/footer.component.scss
@@ -1,5 +1,5 @@
 // PRESET VARIABLES
-@import "../../sass/presets";
+@import "../../../sass/presets";
 
 .footer {
   margin-top: 2.5rem;
