+++ src/app/shared/header/header.component.scss
@@ -1,5 +1,5 @@
 // PRESET VARIABLES
-@import "../../sass/presets";
+@import "../../../sass/presets";
 
 .header {
   margin-top: 1.25rem;
