+++ src/app/shared/navigation/navigation.component.scss
@@ -1,5 +1,5 @@
 // PRESET VARIABLES
-@import "../../sass/presets";
+@import "../../../sass/presets";
 
 .navbar {
   height: 4rem;
