+++ src/app/user/user.component.spec.ts
@@ -2,8 +2,8 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 // COMPONENTS
 import { UserComponent } from './user.component';
-import { HeaderComponent } from '../shared/header.component';
-import { FooterComponent } from '../shared/footer.component';
+import { HeaderComponent } from '../shared/header/header.component';
+import { FooterComponent } from '../shared/footer/footer.component';
 
 describe('UserComponent', () => {
   let component: UserComponent;
